pyproject.toml
[project]
name = "evaluation-parsing"
version = "0.1.0"
description = "Evaluation parsing for Hugging Face models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "claude-agent-sdk>=0.1.19",
    "huggingface-hub>=1.3.2",
    "pyyaml>=6.0.3",
]
